Enhanced Client and ClamdAdmin to mock out the physical connection   to the clamd daemon and processes thus allowing seamless mock tests before   running real tests with the clamd daemon

+++ src/main/java/com/github/jlangch/aviron/admin/ClamdAdmin.java
@@ -26,7 +26,9 @@ import java.io.File;
 import java.io.IOException;
 import java.nio.charset.Charset;
 import java.nio.file.Files;
+import java.util.ArrayList;
 import java.util.List;
+import java.util.concurrent.atomic.AtomicBoolean;
 
 import com.github.jlangch.aviron.ex.AvironException;
 import com.github.jlangch.aviron.ex.NotRunningException;
@@ -65,6 +67,8 @@ public class ClamdAdmin {
      */
     public static String getClamdPID() {
         Shell.validateLinuxOrMacOSX("Admin::getClamdPID");
+
+        if (mocking.get()) return null;
 
         final List<String> pids = Shell.pgrep("clamd");
         return pids.isEmpty() ? null : pids.get(0);
@@ -122,6 +126,8 @@ public class ClamdAdmin {
             throw new IllegalArgumentException("A pid must not be blank!");
         }
 
+        if (mocking.get()) return false;
+
         Shell.validateLinuxOrMacOSX("Admin::isProcessAlive");
 
         return Shell.isProcessAlive(pid);
@@ -143,6 +149,8 @@ public class ClamdAdmin {
      */
     public static List<String> getCpulimitPIDs() {
         Shell.validateLinuxOrMacOSX("Admin::getCpulimitPIDs");
+
+        if (mocking.get()) return new ArrayList<>();
 
         return Shell.pgrep("cpulimit");
     }
@@ -183,18 +191,20 @@ public class ClamdAdmin {
     ) {
         Shell.validateLinuxOrMacOSX("Admin::activateClamdCpuLimit");
 
-        if (StringUtils.isBlank(clamdPID)) {
-            throw new IllegalArgumentException("No Clamd PID!");
-        }
-
         if (limit < 0) {
             throw new IllegalArgumentException(
                     "A limit value must not be negative!");
         }
 
+        if (mocking.get()) return;
+
+        if (StringUtils.isBlank(clamdPID)) {
+            throw new IllegalArgumentException("No Clamd PID!");
+        }
+
         try {
             // /bin/sh -c "nohup /usr/bin/cpulimit -p 1234 -l 50 </dev/null &>/dev/null &"
-            
+
             // run cpulimit as nohup process
             Shell.execCmdBackgroundNohup("cpulimit", "--limit=" + limit, "--pid=" + clamdPID);
         }
@@ -220,6 +230,8 @@ public class ClamdAdmin {
      */
     public static void deactivateClamdCpuLimit(final String clamdPID) {
         Shell.validateLinuxOrMacOSX("Admin::deactivateClamdCpuLimit");
+
+        if (mocking.get()) return;
 
         if (StringUtils.isBlank(clamdPID)) {
             throw new NotRunningException("No Clamd PID!");
@@ -254,6 +266,8 @@ public class ClamdAdmin {
     public static void killClamd() {
         Shell.validateLinuxOrMacOSX("Admin::killClamd");
 
+        if (mocking.get()) return;
+
         final String clamdPID = getClamdPID();
         if (!StringUtils.isBlank(clamdPID)) {
             Shell.kill(Signal.SIGTERM, getClamdPID());
@@ -276,4 +290,16 @@ public class ClamdAdmin {
         return Runtime.getRuntime().availableProcessors();
     }
 
+
+    /**
+     * Enable/disable mocking mode
+     * 
+     * @param mock enable/disable mocking mode
+     */
+    public static void mocking(final boolean mock) {
+        mocking.set(mock);
+    }
+
+
+    private static final AtomicBoolean mocking = new AtomicBoolean(false);
 }
